fixed types created password utils and token utils for better organisation
