+++ api/Business/BusinessObjects/UserBusinessObject.cs
@@ -56,7 +56,7 @@ namespace Business.BusinessObjects
                 {
                     throw new Exception("user does not exist");
                 }
-                if (!IsValidEmail(record.Email!) || record.Name.IsNullOrEmpty() || record.Picture.IsNullOrEmpty() || record.Phone.IsNullOrEmpty() || record.Password!.Length < 8 )
+                if (!EmailUtils.IsValidEmail(record.Email!) || record.Name.IsNullOrEmpty() || record.Picture.IsNullOrEmpty() || record.Phone.IsNullOrEmpty() || record.Password!.Length < 8 )
                 {
                     throw new Exception();
                 }
@@ -66,7 +66,7 @@ namespace Business.BusinessObjects
                     user.Password = record.Password;
                     user.Email = record.Email;
                     user.Picture = record.Picture;
-                    user.Password = EncodePasswordToBase64(record.Password);
+                    user.Password = PasswordUtils.EncodePasswordToBase64(record.Password!);
                     user.Phone = record.Phone;
                     user.Role = record.Role;
                 }
@@ -94,7 +94,8 @@ namespace Business.BusinessObjects
         {
             return await ExecuteOperation(async () =>
             {
-                if (!IsValidEmail(record.Email!) || record.Name.IsNullOrEmpty() || record.Picture.IsNullOrEmpty() || record.Phone.IsNullOrEmpty())
+
+                if (!EmailUtils.IsValidEmail(record.Email!) || record.Name.IsNullOrEmpty() || record.Picture.IsNullOrEmpty() || record.Phone.IsNullOrEmpty())
                 {
                     throw new Exception();
                 }
@@ -105,12 +106,13 @@ namespace Business.BusinessObjects
                    throw new Exception("Email already exists");
                 }
 
-                string password = GenerateRandomPassword();
-                EmailUtils emailUtils = new EmailUtils();
+                string password = PasswordUtils.GenerateRandomPassword();
                 string subject = "Password";
+                
+                EmailUtils emailUtils = new EmailUtils();
                 emailUtils.sendEmail(record.Email!,subject, password);
 
-                record.Password = EncodePasswordToBase64(password);
+                record.Password = PasswordUtils.EncodePasswordToBase64(password);
                 await _genericDataAccessObject.InsertAsync<User>(record);
                 return new CreateUserBusinessModel { Uuid = record.Uuid};
             });
@@ -120,7 +122,7 @@ namespace Business.BusinessObjects
         {
             return await ExecuteOperation(async () =>
             {
-                if (!IsValidEmail(email))
+                if (!EmailUtils.IsValidEmail(email))
                 {
                     throw new Exception("Invalid email");
                 }
@@ -130,11 +132,11 @@ namespace Business.BusinessObjects
                 {
                     throw new Exception("No email found");
                 }
-                if (DecodeFrom64(result.Password!) != password)
+                if (PasswordUtils.DecodeFrom64(result.Password!) != password)
                 {
                     throw new Exception("Invalid password");
                 }
-                string token = CreateToken(result);
+                string token = TokenUtils.CreateAuthenticationToken(result);
 
                 UserTokenAuthentication userToken = await _userDataAccessObject.GetUserTokenByUserId(result.Id);
 
@@ -189,7 +191,7 @@ namespace Business.BusinessObjects
                 }
                 else
                 {
-                    string resetToken = CreateToken(user);
+                    string resetToken = TokenUtils.GeneratePasswordResetToken();
 
                     user.PasswordResetToken = resetToken;
                     user.PasswordResetTokenExpiration = DateTime.UtcNow.AddHours(1); 
@@ -217,94 +219,15 @@ namespace Business.BusinessObjects
                 {
                     throw new Exception("Token expired");
                 }
-                string userPassword=DecodeFrom64(user.Password!);
+                string userPassword= PasswordUtils.DecodeFrom64(user.Password!);
                 if (string.IsNullOrEmpty(newPassword) || newPassword == userPassword) {
                     throw new Exception("Please insert a valid password");
                 }
 
-                user.Password = EncodePasswordToBase64(newPassword);
+                user.Password = PasswordUtils.EncodePasswordToBase64(newPassword);
                 user.PasswordResetTokenExpiration = DateTime.UtcNow;
                 await _genericDataAccessObject.UpdateAsync(user);
             });
-        }
-        private string GenerateRandomPassword(int length = 12)
-        {
-            string lowercaseLetters = "abcdefghijklmnopqrstuvwxyz";
-            string uppercaseLetters = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
-            string digits = "0123456789";
-            string specialCharacters = "!@#$%^&*()_-+=<>?/\\";
-
-            string allCharacters = lowercaseLetters + uppercaseLetters + digits + specialCharacters;
-
-            length = Math.Max(length, 8);
-
-            Random random = new Random();
-            StringBuilder passwordBuilder = new StringBuilder();
-            for (int i = 0; i < length; i++)
-            {
-                int index = random.Next(allCharacters.Length);
-                passwordBuilder.Append(allCharacters[index]);
-            }
-
-            return passwordBuilder.ToString();
-        }
-
-        static bool IsValidEmail(string email)
-        {
-            string pattern = @"^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$";
-            Regex regex = new Regex(pattern);
-
-            return regex.IsMatch(email);
-        }
-
-        private string CreateToken(User user)
-        {
-            IConfiguration configuration = new ConfigurationBuilder()
-            .SetBasePath(Directory.GetCurrentDirectory())
-            .AddJsonFile("appsettings.json")
-            .Build();
-
-            List<Claim> claims = new List<Claim>
-            {
-                new Claim(ClaimTypes.Name, user.Name!),
-            };
-
-            var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(configuration.GetSection("AppSettings:Token").Value!));
-            var creds = new SigningCredentials(key, SecurityAlgorithms.HmacSha256);
-            var token = new JwtSecurityToken(
-                claims:claims,
-                expires:DateTime.Now.AddDays(1),
-                signingCredentials:creds
-                );
-            var jwt = new JwtSecurityTokenHandler().WriteToken(token);
-            return jwt;
-        }
-
-        public static string EncodePasswordToBase64(string password)
-        {
-            try
-            {
-                byte[] encData_byte = new byte[password.Length];
-                encData_byte = System.Text.Encoding.UTF8.GetBytes(password);
-                string encodedData = Convert.ToBase64String(encData_byte);
-                return encodedData;
-            }
-            catch (Exception ex)
-            {
-                throw new Exception("Error in base64Encode" + ex.Message);
-            }
-        }
-
-        public string DecodeFrom64(string encodedData)
-        {
-            System.Text.UTF8Encoding encoder = new System.Text.UTF8Encoding();
-            System.Text.Decoder utf8Decode = encoder.GetDecoder();
-            byte[] todecode_byte = Convert.FromBase64String(encodedData);
-            int charCount = utf8Decode.GetCharCount(todecode_byte, 0, todecode_byte.Length);
-            char[] decoded_char = new char[charCount];
-            utf8Decode.GetChars(todecode_byte, 0, todecode_byte.Length, decoded_char, 0);
-            string result = new String(decoded_char);
-            return result;
         }
     }
 }
