+++ api/Business/Utils/PasswordUtils.cs
@@ -1,0 +1,60 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Business.Utils
+{
+    internal class PasswordUtils
+    {
+        public static string EncodePasswordToBase64(string password)
+        {
+            try
+            {
+                byte[] encData_byte = new byte[password.Length];
+                encData_byte = System.Text.Encoding.UTF8.GetBytes(password);
+                string encodedData = Convert.ToBase64String(encData_byte);
+                return encodedData;
+            }
+            catch (Exception ex)
+            {
+                throw new Exception("Error in base64Encode" + ex.Message);
+            }
+        }
+
+        public static string DecodeFrom64(string encodedData)
+        {
+            System.Text.UTF8Encoding encoder = new System.Text.UTF8Encoding();
+            System.Text.Decoder utf8Decode = encoder.GetDecoder();
+            byte[] todecode_byte = Convert.FromBase64String(encodedData);
+            int charCount = utf8Decode.GetCharCount(todecode_byte, 0, todecode_byte.Length);
+            char[] decoded_char = new char[charCount];
+            utf8Decode.GetChars(todecode_byte, 0, todecode_byte.Length, decoded_char, 0);
+            string result = new String(decoded_char);
+            return result;
+        }
+
+        public static string GenerateRandomPassword(int length = 12)
+        {
+            string lowercaseLetters = "abcdefghijklmnopqrstuvwxyz";
+            string uppercaseLetters = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
+            string digits = "0123456789";
+            string specialCharacters = "!@#$%^&*()_-+=<>?/\\";
+
+            string allCharacters = lowercaseLetters + uppercaseLetters + digits + specialCharacters;
+
+            length = Math.Max(length, 8);
+
+            Random random = new Random();
+            StringBuilder passwordBuilder = new StringBuilder();
+            for (int i = 0; i < length; i++)
+            {
+                int index = random.Next(allCharacters.Length);
+                passwordBuilder.Append(allCharacters[index]);
+            }
+
+            return passwordBuilder.ToString();
+        }
+    }
+}
