+++ api/DataAccess/DataAccessObjects/UserDataAccessObject.cs
@@ -47,7 +47,7 @@ namespace DataAccess.DataAccessObjects
 
         public async Task<User> GetUserByEmail(string email)
         {
-            var result = await _context.Set<User>().Where(x => x.Email == email)
+            List<User> result = await _context.Set<User>().Where(x => x.Email == email)
                 .ToListAsync();
 
             return result.Where(x => !x.IsDeleted).SingleOrDefault()!;
@@ -55,18 +55,18 @@ namespace DataAccess.DataAccessObjects
 
         public async Task<UserTokenAuthentication> GetUserTokenByUserId(long id)
         {
-            var result = await _context.Set<UserTokenAuthentication>().Where(x => x.UserId == id).ToListAsync();
+            List<UserTokenAuthentication> result = await _context.Set<UserTokenAuthentication>().Where(x => x.UserId == id).ToListAsync();
             return result.Where(x => !x.IsDeleted).SingleOrDefault()!;
         }
 
         public async Task<UserTokenAuthentication> GetTokenUuidByToken(string token)
         {
-            var result = await _context.Set<UserTokenAuthentication>().Where(x => x.Token == token).ToListAsync();
+            List<UserTokenAuthentication> result = await _context.Set<UserTokenAuthentication>().Where(x => x.Token == token).ToListAsync();
             return result.Where(x => !x.IsDeleted).SingleOrDefault()!;
         }
         public async Task<User> GetUserByPasswordResetToken(string token)
         {
-            var result = await _context.Set<User>().Where(x=>x.PasswordResetToken == token).ToListAsync();
+            List<User> result = await _context.Set<User>().Where(x=>x.PasswordResetToken == token).ToListAsync();
             return result.Where(x => !x.IsDeleted).SingleOrDefault()!;
         }
     }
